+++ WordHunt.Data/Entities/GameField.cs
@@ -7,13 +7,13 @@ namespace WordHunt.Data.Entities
 {
     public class GameField
     {
-        public long Id { get; set; }
-        public long GameId { get; set; }
+        public int Id { get; set; }
+        public int GameId { get; set; }
         public string Word { get; set; }
         public FieldType Type { get; set; }
-        public long? TeamId { get; set; }
+        public int? TeamId { get; set; }
         public bool Checked { get; set; }
-        public long? CheckedByTeamId { get; set; }
+        public int? CheckedByTeamId { get; set; }
         public int ColumnIndex { get; set; }
         public int RowIndex { get; set; }
     }
