+++ WordHunt.DataInterfaces/Words/Request/WordCreateRequest.cs
@@ -7,7 +7,7 @@ namespace WordHunt.DataInterfaces.Words.Request
     public class WordCreateRequest
     {
         public string Value { get; set; }
-        public long LanguageId { get; set; }
-        public long? CategoryId { get; set; }
+        public int LanguageId { get; set; }
+        public int? CategoryId { get; set; }
     }
 }
