+++ WordHunt.DataInterfaces/Words/Request/WordListGetRequest.cs
@@ -6,8 +6,8 @@ namespace WordHunt.DataInterfaces.Words.Request
 {
     public class WordListGetRequest
     {
-        public long LanguageId { get; set; }
-        public long? CategoryId { get; set; }
+        public int LanguageId { get; set; }
+        public int? CategoryId { get; set; }
         public int PageSize { get; set; }
         public int Page { get; set; }
         public bool OrderByDesc { get; set; }
