+++ WordHunt.DataInterfaces/Words/Request/WordUpdateRequest.cs
@@ -8,6 +8,6 @@ namespace WordHunt.DataInterfaces.Words.Request
     {
         public int Id { get; set; }
         public string Value { get; set; }
-        public long? CategoryId { get; set; }
+        public int? CategoryId { get; set; }
     }
 }
